refactored display code into classes
